+++ tests/Feature/Fields/FieldArrangementTest.php
@@ -2,7 +2,12 @@
 
 namespace Thinktomorrow\Chief\Tests\Feature\Management;
 
+use Thinktomorrow\Chief\Fields\FieldArrangement;
+use Thinktomorrow\Chief\Fields\Fields;
+use Thinktomorrow\Chief\Fields\FieldsTab;
+use Thinktomorrow\Chief\Fields\RemainingFieldsTab;
 use Thinktomorrow\Chief\Fields\Types\Field;
+use Thinktomorrow\Chief\Fields\Types\InputField;
 use Thinktomorrow\Chief\Management\Register;
 use Thinktomorrow\Chief\Tests\Feature\Management\Fakes\ManagedModelFake;
 use Thinktomorrow\Chief\Tests\Feature\Management\Fakes\ManagedModelFakeTranslation;
@@ -28,6 +33,47 @@ class FieldArrangementTest extends TestCase
 
         $this->model = ManagedModelFake::create(['title' => 'Foobar', 'custom_column' => 'custom']);
         $this->manager = (new ManagerFake(app(Register::class)->first()))->manage($this->model);
+    }
+
+    /** @test */
+    public function a_tab_can_be_added()
+    {
+        $arrangement = new FieldArrangement(new Fields([]), [new FieldsTab('first-tab')]);
+
+        $this->assertCount(1, $arrangement->tabs());
+        $this->assertCount(2, $arrangement->addTab(new FieldsTab('second-tab'))->tabs());
+    }
+
+    /** @test */
+    public function a_tab_can_be_added_at_specific_position()
+    {
+        $arrangement = new FieldArrangement(new Fields([]), [new FieldsTab('first-tab')]);
+
+        $arrangement = $arrangement->addTab(new FieldsTab('second-tab'), 0);
+
+        $this->assertEquals('second-tab', $arrangement->tabs()[0]->title());
+    }
+
+    /** @test */
+    public function adding_a_tab_is_an_immutable_action()
+    {
+        $arrangement = new FieldArrangement(new Fields([]), [new FieldsTab('first-tab')]);
+
+        $this->assertCount(2, $arrangement->addTab(new FieldsTab('second-tab'))->tabs());
+        $this->assertCount(1, $arrangement->tabs());
+    }
+
+    /** @test */
+    public function remaining_field_is_automatically_placed_in_last_tab()
+    {
+        $arrangement = new FieldArrangement(new Fields([
+            InputField::make('first-input-field'), InputField::make('second-input-field')
+        ]), [new FieldsTab('first-tab', ['first-input-field'])]);
+
+        $arrangement = $arrangement->addTab(new RemainingFieldsTab('second-tab', []));
+
+        $this->assertCount(2, $arrangement->tabs());
+        $this->assertCount(2, $arrangement->fields());
     }
 
     /** @test */